rename ui actions to dialog
