+++ src/electron/dialog/dialog.js
